requirements.txt
alembic==1.4.1
aniso8601==6.0.0
appdirs==1.4.3
astroid==2.2.5
attrs==19.3.0
autopep8==1.5.1
awscli==1.18.42
botocore==1.15.42
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.3
distlib==0.3.0
docutils==0.15.2
ecdsa==0.13.2
filelock==3.0.12
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.17.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.5
lazy-object-proxy==1.4.0
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.18.1
packaging==20.3
pandas==0.25.3
pipenv==2018.11.26
pluggy==0.13.1
psycopg2==2.8.4
psycopg2-binary==2.8.2
py==1.8.1
pyasn1==0.4.8
pycodestyle==2.5.0
pycryptodome==3.6.6
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2019.1
PyYAML==5.3.1
requests==2.22.0
rsa==3.4.2
s3transfer==0.3.3
six==1.12.0
SQLAlchemy==1.3.4
typed-ast==1.3.5
urllib3==1.25.6
virtualenv==20.0.13
virtualenv-clone==0.5.3
wcwidth==0.1.9
Werkzeug==0.15.2
wrapt==1.11.1
xlrd==1.2.0
zipp==3.1.0
